[build-scripts] Fix to allow an interrupted oe build to continue.
Fixes OXT-494

+++ build-scripts/oe/build.sh
@@ -29,14 +29,18 @@ BRANCH=%BRANCH%
 SUBNET_PREFIX=%SUBNET_PREFIX%
 ALL_BUILDS_SUBDIR_NAME=%ALL_BUILDS_SUBDIR_NAME%
 
-mkdir $BUILD_DIR
+mkdir -p $BUILD_DIR
 cd $BUILD_DIR
 
-git clone -b $BRANCH git://${SUBNET_PREFIX}.${IP_C}.1/${BUILD_USER}/openxt.git
+if [ ! -d openxt ] ; then
+    git clone -b ${BRANCH} git://${SUBNET_PREFIX}.${IP_C}.1/${BUILD_USER}/openxt.git
+fi
 
 cd openxt
-cp example-config .config
-cat >>.config <<EOF
+
+if [ ! -e .config ] ; then
+    cp example-config .config
+    cat >>.config <<EOF
 BRANCH="${BRANCH}"
 OPENXT_GIT_MIRROR="${SUBNET_PREFIX}.${IP_C}.1/${BUILD_USER}"
 OPENXT_GIT_PROTOCOL="git"
@@ -45,6 +49,7 @@ REPO_DEV_CACERT="/home/build/certs/dev-cacert.pem"
 REPO_DEV_SIGNING_CERT="/home/build/certs/dev-cacert.pem"
 REPO_DEV_SIGNING_KEY="/home/build/certs/dev-cakey.pem"
 EOF
+fi
 
 ./do_build.sh | tee build.log
